Move zip code to persistence.zip package.
