+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/persistence/zip/IZipFileReader.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.otter.common.domain.resourcecontainer.projectimportexport
+package org.wycliffeassociates.otter.common.persistence.zip
 
 import java.io.BufferedReader
 import java.io.File
